将留言板相关合并在一起
