+++ lib/board/bloc/board_home/board_home_state.dart
