+++ lib/board/bloc/topic_detail/topic_detail_state.dart
