+++ lib/board/bloc/topic_edit/topic_edit_bloc.dart
@@ -1,7 +1,7 @@
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:smarthome/models/board.dart';
-import 'package:smarthome/repositories/board_repository.dart';
+import 'package:smarthome/board/board_repository.dart';
 import 'package:smarthome/utils/exceptions.dart';
 
 part 'topic_edit_event.dart';
