+++ lib/board/board.dart
@@ -1,0 +1,4 @@
+export 'view/home_page.dart';
+export 'view/topic_detail_page.dart';
+export 'board_repository.dart';
+export 'bloc/blocs.dart';
