+++ lib/board/graphql/mutations/closeTopic.dart
