+++ lib/board/graphql/mutations/deleteTopic.dart
