+++ lib/board/graphql/queries/queries.dart
