+++ lib/board/graphql/queries/topic_detail.dart
