+++ lib/board/view/comment_edit_page.dart
@@ -2,62 +2,59 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import 'package:smarthome/blocs/core/blocs.dart';
-import 'package:smarthome/blocs/board/blocs.dart';
+import 'package:smarthome/board/bloc/blocs.dart';
 import 'package:smarthome/models/board.dart';
-import 'package:smarthome/pages/board/widgets/topic_item.dart';
+import 'package:smarthome/board/view/widgets/comment_item.dart';
 import 'package:smarthome/utils/show_snack_bar.dart';
 
-class TopicEditPage extends StatefulWidget {
-  final bool? isEditing;
-  final Topic? topic;
+class CommentEditPage extends StatefulWidget {
+  final bool isEditing;
+  final Comment? comment;
 
-  const TopicEditPage({
+  const CommentEditPage({
     Key? key,
-    this.isEditing,
-    this.topic,
+    required this.isEditing,
+    this.comment,
   }) : super(key: key);
 
   @override
-  _TopicEditPageState createState() => _TopicEditPageState();
+  _CommentEditPageState createState() => _CommentEditPageState();
 }
 
-class _TopicEditPageState extends State<TopicEditPage> {
-  final _titleController = TextEditingController();
-  final _descriptionController = TextEditingController();
+class _CommentEditPageState extends State<CommentEditPage> {
+  final _bodyController = TextEditingController();
   final _formKey = GlobalKey<FormState>();
 
   @override
   void initState() {
     super.initState();
-    if (widget.isEditing!) {
-      _titleController.text = widget.topic!.title!;
-      _descriptionController.text = widget.topic!.description!;
+    if (widget.isEditing) {
+      _bodyController.text = widget.comment!.body!;
     }
   }
 
   @override
   void dispose() {
-    _titleController.dispose();
-    _descriptionController.dispose();
+    _bodyController.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     final List<String> tabs = ['编辑', '预览'];
-    return BlocListener<TopicEditBloc, TopicEditState>(
+    return BlocListener<CommentEditBloc, CommentEditState>(
       listener: (context, state) {
-        if (state is TopicAddSuccess || state is TopicUpdateSuccess) {
+        if (state is CommentAddSuccess || state is CommentUpdateSuccess) {
           Navigator.of(context).pop();
-          if (widget.isEditing!) {
-            showInfoSnackBar('话题编辑成功');
+          if (widget.isEditing) {
+            showInfoSnackBar('评论编辑成功');
           } else {
             BlocProvider.of<BoardHomeBloc>(context)
                 .add(BoardHomeFetched(cache: false));
-            showInfoSnackBar('话题添加成功');
+            showInfoSnackBar('评论添加成功');
           }
         }
-        if (state is TopicFailure) {
+        if (state is CommentFailure) {
           showErrorSnackBar(state.message);
         }
       },
@@ -65,7 +62,7 @@ class _TopicEditPageState extends State<TopicEditPage> {
         length: tabs.length,
         child: Scaffold(
           appBar: AppBar(
-            title: widget.isEditing! ? Text('编辑话题') : Text('新话题'),
+            title: widget.isEditing ? Text('编辑评论') : Text('新评论'),
             actions: [
               Tooltip(
                 message: '提交',
@@ -73,17 +70,11 @@ class _TopicEditPageState extends State<TopicEditPage> {
                   icon: Icon(Icons.send),
                   onPressed: () {
                     if (_formKey.currentState!.validate()) {
-                      if (widget.isEditing!) {
-                        BlocProvider.of<TopicEditBloc>(context)
-                            .add(TopicUpdated(
-                          id: widget.topic!.id,
-                          title: _titleController.text,
-                          description: _descriptionController.text,
-                        ));
-                      } else {
-                        BlocProvider.of<TopicEditBloc>(context).add(TopicAdded(
-                          title: _titleController.text,
-                          description: _descriptionController.text,
+                      if (widget.isEditing) {
+                        BlocProvider.of<CommentEditBloc>(context)
+                            .add(CommentUpdated(
+                          id: widget.comment!.id,
+                          body: _bodyController.text,
                         ));
                       }
                       showInfoSnackBar('正在提交...', duration: 1);
@@ -100,14 +91,12 @@ class _TopicEditPageState extends State<TopicEditPage> {
             children: [
               _EditPage(
                 isEditing: widget.isEditing,
-                topic: widget.topic,
+                comment: widget.comment,
                 formKey: _formKey,
-                titleController: _titleController,
-                descriptionController: _descriptionController,
+                bodyController: _bodyController,
               ),
               _PreviewPage(
-                titleController: _titleController,
-                descriptionController: _descriptionController,
+                bodyController: _bodyController,
               ),
             ],
           ),
@@ -118,18 +107,16 @@ class _TopicEditPageState extends State<TopicEditPage> {
 }
 
 class _EditPage extends StatelessWidget {
-  final bool? isEditing;
-  final Topic? topic;
-  final TextEditingController titleController;
-  final TextEditingController descriptionController;
+  final bool isEditing;
+  final Comment? comment;
+  final TextEditingController? bodyController;
   final GlobalKey<FormState> formKey;
 
   const _EditPage({
     Key? key,
     required this.isEditing,
-    this.topic,
-    required this.titleController,
-    required this.descriptionController,
+    this.comment,
+    this.bodyController,
     required this.formKey,
   }) : super(key: key);
 
@@ -143,22 +130,7 @@ class _EditPage extends StatelessWidget {
           child: Column(
             children: [
               TextFormField(
-                controller: titleController,
-                decoration: InputDecoration(
-                  border: InputBorder.none,
-                  hintText: '添加标题',
-                  hintStyle: TextStyle(fontSize: 18),
-                ),
-                validator: (value) {
-                  if (value!.isEmpty) {
-                    return '请填写标题';
-                  }
-                  return null;
-                },
-                autovalidateMode: AutovalidateMode.onUserInteraction,
-              ),
-              TextFormField(
-                controller: descriptionController,
+                controller: bodyController,
                 maxLines: null,
                 decoration: InputDecoration(
                   border: InputBorder.none,
@@ -182,13 +154,11 @@ class _EditPage extends StatelessWidget {
 }
 
 class _PreviewPage extends StatefulWidget {
-  final TextEditingController titleController;
-  final TextEditingController descriptionController;
+  final TextEditingController? bodyController;
 
   const _PreviewPage({
     Key? key,
-    required this.titleController,
-    required this.descriptionController,
+    this.bodyController,
   }) : super(key: key);
 
   @override
@@ -199,14 +169,12 @@ class __PreviewPageState extends State<_PreviewPage> {
   @override
   void initState() {
     super.initState();
-    widget.titleController.addListener(_onChanged);
-    widget.descriptionController.addListener(_onChanged);
+    widget.bodyController!.addListener(_onChanged);
   }
 
   @override
   void dispose() {
-    widget.titleController.removeListener(_onChanged);
-    widget.descriptionController.removeListener(_onChanged);
+    widget.bodyController!.removeListener(_onChanged);
     super.dispose();
   }
 
@@ -219,16 +187,15 @@ class __PreviewPageState extends State<_PreviewPage> {
     final loginUser =
         context.select((AppPreferencesBloc b) => b.state.loginUser);
     return SingleChildScrollView(
-      child: TopicItem(
-        topic: Topic(
+      child: CommentItem(
+        comment: Comment(
           id: '',
           user: loginUser,
-          title: widget.titleController.text,
-          description: widget.descriptionController.text,
+          body: widget.bodyController!.text,
           createdAt: DateTime.now(),
           editedAt: DateTime.now(),
         ),
-        showBody: true,
+        showMenu: false,
       ),
     );
   }
