+++ lib/board/view/home_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'package:smarthome/blocs/board/blocs.dart';
+import 'package:smarthome/board/bloc/blocs.dart';
+import 'package:smarthome/board/view/topic_edit_page.dart';
+import 'package:smarthome/board/view/widgets/topic_item.dart';
 import 'package:smarthome/models/app_tab.dart';
-import 'package:smarthome/pages/board/topic_edit_page.dart';
-import 'package:smarthome/pages/board/widgets/topic_item.dart';
-import 'package:smarthome/repositories/board_repository.dart';
+import 'package:smarthome/board/board_repository.dart';
 import 'package:smarthome/widgets/center_loading_indicator.dart';
 import 'package:smarthome/widgets/error_message_button.dart';
 import 'package:smarthome/widgets/home_page.dart';
