+++ lib/board/view/topic_detail_page.dart
@@ -4,14 +4,13 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
 
 import 'package:smarthome/blocs/core/blocs.dart';
-import 'package:smarthome/blocs/board/blocs.dart';
-import 'package:smarthome/blocs/board/topic_detail/topic_detail_bloc.dart';
+import 'package:smarthome/board/bloc/blocs.dart';
+import 'package:smarthome/board/view/topic_edit_page.dart';
+import 'package:smarthome/board/view/widgets/add_comment_bar.dart';
+import 'package:smarthome/board/view/widgets/comment_item.dart';
+import 'package:smarthome/board/view/widgets/topic_item.dart';
 import 'package:smarthome/models/models.dart';
-import 'package:smarthome/pages/board/topic_edit_page.dart';
-import 'package:smarthome/pages/board/widgets/add_comment_bar.dart';
-import 'package:smarthome/pages/board/widgets/comment_item.dart';
-import 'package:smarthome/pages/board/widgets/topic_item.dart';
-import 'package:smarthome/repositories/board_repository.dart';
+import 'package:smarthome/board/board_repository.dart';
 import 'package:smarthome/widgets/center_loading_indicator.dart';
 import 'package:smarthome/widgets/error_message_button.dart';
 import 'package:smarthome/utils/show_snack_bar.dart';
