+++ lib/board/view/widgets/comment_item.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
-import 'package:smarthome/blocs/board/blocs.dart';
+import 'package:smarthome/board/bloc/blocs.dart';
 import 'package:smarthome/blocs/core/blocs.dart';
+import 'package:smarthome/board/view/comment_edit_page.dart';
+import 'package:smarthome/board/view/widgets/item_title.dart';
 import 'package:smarthome/models/models.dart';
-import 'package:smarthome/pages/board/comment_edit_page.dart';
-import 'package:smarthome/pages/board/widgets/item_title.dart';
-import 'package:smarthome/repositories/board_repository.dart';
+import 'package:smarthome/board/board_repository.dart';
 import 'package:smarthome/utils/show_snack_bar.dart';
 
 class CommentItem extends StatelessWidget {
