+++ lib/board/view/widgets/topic_item.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
+import 'package:smarthome/board/view/topic_detail_page.dart';
+import 'package:smarthome/board/view/widgets/item_title.dart';
 import 'package:smarthome/models/board.dart';
-import 'package:smarthome/pages/board/topic_detail_page.dart';
-import 'package:smarthome/pages/board/widgets/item_title.dart';
 import 'package:smarthome/routers/delegate.dart';
 
 class TopicItem extends StatelessWidget {
