+++ lib/main.dart
@@ -8,8 +8,9 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:intl/intl.dart';
 import 'package:smarthome/app_config.dart';
 import 'package:smarthome/blocs/core/blocs.dart';
-import 'package:smarthome/blocs/board/blocs.dart';
+import 'package:smarthome/board/bloc/blocs.dart';
 import 'package:smarthome/blocs/storage/blocs.dart';
+import 'package:smarthome/board/board_repository.dart';
 import 'package:smarthome/models/grobal_keys.dart';
 import 'package:smarthome/repositories/iot_repository.dart';
 import 'package:smarthome/repositories/push_repository.dart';
