+++ lib/models/app_tab.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:smarthome/board/view/home_page.dart';
 import 'package:smarthome/pages/blog/home_page.dart';
-import 'package:smarthome/pages/board/home_page.dart';
 import 'package:smarthome/pages/iot/home_page.dart';
 import 'package:smarthome/pages/storage/home_page.dart';
