+++ lib/repositories/repositories.dart
@@ -1,4 +1,3 @@
-export 'board_repository.dart';
 export 'graphql_api_client.dart';
 export 'storage_repository.dart';
 export 'user_repository.dart';
