+++ lib/routers/delegate.dart
@@ -8,11 +8,10 @@ import 'package:logging/logging.dart';
 import 'package:quick_actions/quick_actions.dart';
 import 'package:smarthome/app_config.dart';
 import 'package:smarthome/blocs/core/blocs.dart';
+import 'package:smarthome/board/board.dart';
 import 'package:smarthome/models/grobal_keys.dart';
 import 'package:smarthome/models/models.dart';
 import 'package:smarthome/pages/blog/home_page.dart';
-import 'package:smarthome/pages/board/home_page.dart';
-import 'package:smarthome/pages/board/topic_detail_page.dart';
 import 'package:smarthome/pages/iot/home_page.dart';
 import 'package:smarthome/pages/login_page.dart';
 import 'package:smarthome/pages/settings/blog/settings_page.dart';
